+++ MSGraph/functions/exchange/mail/Remove-MgaMailMessage.ps1
@@ -105,7 +105,7 @@
                     "Force"        = $true
                     "FunctionName" = $MyInvocation.MyCommand
                 }
-                $null = Invoke-MgaDeleteMethod @invokeParam
+                $null = Invoke-MgaRestMethodDelete @invokeParam
                 if ($PassThru) {
                     $messageItem.InputObject
                 }